@@ -8,19 +8,33 @@ using MediaBrowser.Model.Serialization;
 
 namespace Jellyfin.Plugin.Keycloak
 {
+    /// <summary>
+    /// Keycloak Plugin.
+    /// </summary>
     public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
     {
-        public override string Name => "Keycloak-Auth";
-
-        public override Guid Id => Guid.Parse("40886866-b3dd-4d6a-bf9b-25c83e6c3d10");
-
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Plugin"/> class.
+        /// </summary>
+        /// <param name="applicationPaths">Instance of the <see cref="IApplicationPaths"/> interface.</param>
+        /// <param name="xmlSerializer">Instance of the <see cref="IXmlSerializer"/>interface.</param>
         public Plugin(IApplicationPaths applicationPaths, IXmlSerializer xmlSerializer) : base(applicationPaths, xmlSerializer)
         {
             Instance = this;
         }
 
-        public static Plugin Instance { get; private set; }
+        /// <summary>
+        /// Gets the plugin instance.
+        /// </summary>
+        public static Plugin Instance { get; private set; } = null!;
 
+        /// <inheritdoc />
+        public override string Name => "Keycloak-Auth";
+
+        /// <inheritdoc />
+        public override Guid Id => Guid.Parse("40886866-b3dd-4d6a-bf9b-25c83e6c3d10");
+
+        /// <inheritdoc />
         public IEnumerable<PluginPageInfo> GetPages()
         {
             return new[]
@@ -28,7 +42,7 @@ namespace Jellyfin.Plugin.Keycloak
                 new PluginPageInfo
                 {
                     Name = this.Name,
-                    EmbeddedResourcePath = string.Format("{0}.Configuration.configPage.html", GetType().Namespace)
+                    EmbeddedResourcePath = $"{GetType().Namespace}.Configuration.configPage.html"
                 }
             };
         }
